InstallerGUI/requirements.txt
future==0.15.2
pefile==2016.3.28
Pillow==3.4.2
PyInstaller==3.2
pypiwin32==219
